+++ src/repositories/postRepository.js
@@ -35,8 +35,8 @@ async function getAllPosts() {
       p."description",
       p."createdAt" AS "postCreatedAt",
       p."editedAt",
-      COUNT("likesPosts"."id") as "likes",
-      COUNT(comments."postId") AS "commentsCount",
+      COUNT("likesPosts"."id")::int as "likes",
+      COUNT(comments."postId")::int AS "commentsCount",
       p."urlTitle",
       p."urlDescription",
       p."urlImage"
